// src/app/api/models/paginated-response-post-model.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PostModel } from '../models/post-model';

/**
 * 分頁 API 回傳格式
 */
export interface PaginatedResponsePostModel {
  data: Array<PostModel>;
  page: number;
  pageSize: number;
  total: number;
  totalPages: number;
}